backend/app/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
google-genai
opencv-python-headless
python-dotenv
